Частично решил проблему с долгой отправкой сообщений при спаме
Текущая проблема:
пока не понял как колбэком вернуть сообщению его Id из-за измененной структуры.
Как это работает расскажу в понедельник,  это пока не так важно

+++ Web/Controllers/HomeController.cs
@@ -47,7 +47,7 @@ namespace Web.Controllers
         }
 
         [HttpPost]
-        public async Task<JsonResult> SendMessage(string text, int groupId, int hash)
+        public async Task<JsonResult> SendMessage(string text, int groupId, long hash)
         {
             var msg = new MessageWCF
             {
